+++ tests/unit/Classes/OtherStringIdentifier.php
@@ -1,0 +1,26 @@
+<?php
+declare(strict_types=1);
+
+namespace FireMidge\Tests\Identifier\Unit\Classes;
+
+use FireMidge\Identifier\Identifier;
+
+class OtherStringIdentifier implements Identifier
+{
+    public function __construct(private string $value) {}
+
+    public function toString() : string
+    {
+        return $this->value;
+    }
+
+    public function __toString() : string
+    {
+        return $this->value;
+    }
+
+    public function isEqualTo(Identifier $id) : bool
+    {
+        return false;
+    }
+}